# openapi.yaml
openapi: "3.1.0"
info:
  title: Stellar Wrap API
  version: "1.0.0"
  description: |
    API routes exposed by the Stellar Wrap Next.js application.

    Public endpoints:
    - **GET /api/wrapped** — returns aggregated on-chain statistics for a Stellar address.
    - **GET /api/og** — returns a 1200 × 1200 PNG share card (served from Vercel Edge Runtime).

    Notification endpoints (push + email subscriptions, GDPR data deletion):
    - **POST /api/notifications/subscribe** — store a browser push subscription.
    - **POST /api/notifications/subscribe-email** — start an email subscription (pending confirmation).
    - **GET /api/notifications/confirm-email** — confirm a pending email subscription.
    - **POST /api/notifications/unsubscribe** — unsubscribe by token or by wallet + channel.
    - **POST /api/notifications/dispatch** — triggered by a Vercel cron; fans out due notifications.
    - **GET /api/notifications/preferences/{wallet}** — read a subscription record.
    - **PUT /api/notifications/preferences/{wallet}** — update a subscription record.
    - **DELETE /api/notifications/data/{wallet}** — GDPR data deletion request.
servers:
  - url: https://stellar-wrap.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
security: []
paths:
...
  /api/notifications/subscribe:
    post:
      operationId: subscribePush
      summary: Store a browser push subscription for a wallet
      description: |
        Persists a Web Push subscription for a Stellar wallet address together
        with the wrap periods the subscriber wants to be notified about.

        Requires a valid `PushSubscriptionJSON` object from the
        browser's Push API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscribeRequest"
            example:
              walletAddress: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
              subscription:
                endpoint: "https://fcm.googleapis.com/fcm/send/abc123"
                expirationTime: null
                keys:
                  p256dh: "BHcq...p256dh"
                  auth: "Qxi...auth"
              periods:
                weekly: false
                monthly: true
                yearly: false
      responses:
        "200":
          description: Push subscription stored successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              example:
                ok: true
        "400":
          description: Invalid wallet address or missing push subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidWallet:
                  value:
                    error: "Invalid wallet address"
                invalidSubscription:
                  value:
                    error: "Invalid push subscription"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
  /api/notifications/subscribe-email:
    post:
      operationId: subscribeEmail
      summary: Start an email subscription (pending confirmation)
      description: |
        Validates an email address and wallet, writes a pending email
        subscription to KV, and sends a confirmation email with a
        confirmation link. The email is only active after the link is visited
        (`GET /api/notifications/confirm-email`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSubscribeRequest"
            example:
              walletAddress: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
              email: "alice@example.com"
              periods:
                weekly: true
                monthly: true
                yearly: false
      responses:
        "200":
          description: Confirmation email sent; subscription is pending.
          content:
            application/json:
              schema:
                type: object
                required: [ok, status]
                properties:
                  ok:
                    type: boolean
                    example: true
                  status:
                    type: string
                    enum: [pending]
                    example: "pending"
              example:
                ok: true
                status: "pending"
        "400":
          description: Invalid wallet address or invalid email address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidWallet:
                  value:
                    error: "Invalid wallet address"
                invalidEmail:
                  value:
                    error: "Invalid email address"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
...
  /api/notifications/unsubscribe:
    post:
      operationId: unsubscribe
      summary: Unsubscribe from notifications
      description: |
        Two modes, mutually exclusive:
        - `{ "token": "..." }` — one-click unsubscribe using the token from the
          footer of a notification email.
        - `{ "walletAddress": "...", "channel": "push" | "email" }` — disable a
          single channel from the preference page.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeRequest"
            examples:
              byToken:
                value:
                  token: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
              byChannel:
                value:
                  walletAddress: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
                  channel: "email"
      responses:
        "200":
          description: Successfully unsubscribed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              example:
                ok: true
        "400":
          description: Missing token or walletAddress+channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Provide either token or walletAddress+channel"
        "404":
          description: Token not found or no subscription for the wallet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                tokenNotFound:
                  value:
                    error: "Token not found"
                noSubscription:
                  value:
                    error: "No subscription found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
...
  /api/notifications/preferences/{wallet}:
    get:
      operationId: getNotificationPreferences
      summary: Read a subscription record for a wallet
      description: |
        Returns the full `SubscriptionRecord` (push/email channels, period
        preferences, consent flags) for the wallet, or a 404 when no record
        exists.
      parameters:
        - name: wallet
          in: path
          required: true
          description: Stellar public key to read.
          schema:
            type: string
            pattern: "^G[A-Z2-7]{55}$"
            example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
      responses:
        "200":
          description: Subscription record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionRecord"
              example:
                walletAddress: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
                email:
                  address: "alice@example.com"
                  status: "active"
                  confirmationToken: ""
                  unsubscribeToken: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
                  periods:
                    weekly: true
                    monthly: true
                    yearly: false
                  createdAt: "2026-08-01T10:00:00.000Z"
                consentGiven: true
                consentTimestamp: "2026-08-01T10:00:00.000Z"
        "400":
          description: Invalid wallet address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid wallet address"
        "404":
          description: No subscription found for the wallet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "No subscription found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
    put:
      operationId: updateNotificationPreferences
      summary: Update a subscription record for a wallet
      description: |
        Merges only the allowed fields (`push`, `email`, `consentGiven`) into
        the existing record. The `walletAddress` is never overwritten. Returns
        the updated record, or a 404 when no record exists.
      parameters:
        - name: wallet
          in: path
          required: true
          description: Stellar public key to update.
          schema:
            type: string
            pattern: "^G[A-Z2-7]{55}$"
            example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Partial fields to merge into the existing record.
              properties:
                consentGiven:
                  type: boolean
                  example: true
              example:
                consentGiven: true
      responses:
        "200":
          description: Updated subscription record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionRecord"
        "400":
          description: Invalid wallet address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Invalid wallet address"
        "404":
          description: No subscription found for the wallet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "No subscription found"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
components:
  securitySchemes:
    cronSecretAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token matching the `CRON_SECRET` environment variable. Set by the
        Vercel cron job that calls `POST /api/notifications/dispatch`.
  schemas:
...
    PeriodPrefs:
      type: object
      required: [weekly, monthly, yearly]
      properties:
        weekly:
          type: boolean
          example: true
        monthly:
          type: boolean
          example: true
        yearly:
          type: boolean
          example: false
    PushSubscription:
      type: object
      description: A Web Push subscription as produced by the browser Push API.
      required: [endpoint, keys]
      properties:
        endpoint:
          type: string
          format: uri
          description: Push service endpoint URL.
          example: "https://fcm.googleapis.com/fcm/send/abc123"
        expirationTime:
          type: [string, "null"]
          format: date-time
          description: Subscription expiration time, or null if it never expires.
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
              description: Base64url-encoded ECDH P-256 public key.
              example: "BHcq...p256dh"
            auth:
              type: string
              description: Base64url-encoded authentication secret.
              example: "Qxi...auth"
    PushChannel:
      type: object
      required: [subscription, periods, createdAt]
      properties:
        subscription:
          $ref: "#/components/schemas/PushSubscription"
        periods:
          $ref: "#/components/schemas/PeriodPrefs"
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 creation time.
    EmailChannel:
      type: object
      required: [address, status, confirmationToken, unsubscribeToken, periods, createdAt]
      properties:
        address:
          type: string
          format: email
          example: "alice@example.com"
        status:
          type: string
          enum: [pending, active]
          example: "active"
        confirmationToken:
          type: string
          description: Set while pending; cleared on confirmation.
          example: ""
        unsubscribeToken:
          type: string
          description: Persistent token used by the unsubscribe link.
          example: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
        periods:
          $ref: "#/components/schemas/PeriodPrefs"
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 creation time.
    SubscriptionRecord:
      type: object
      required: [walletAddress, consentGiven, consentTimestamp]
      properties:
        walletAddress:
          type: string
          example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
        push:
          $ref: "#/components/schemas/PushChannel"
        email:
          $ref: "#/components/schemas/EmailChannel"
        consentGiven:
          type: boolean
          example: true
        consentTimestamp:
          type: string
          format: date-time
          description: ISO-8601 consent time.
        deletionRequested:
          type: [string, "null"]
          format: date-time
          description: |
            ISO-8601 timestamp set when the user has requested data deletion.
            Present only after a GDPR deletion request.
    PushSubscribeRequest:
      type: object
      required: [walletAddress, subscription, periods]
      properties:
        walletAddress:
          type: string
          pattern: "^G[A-Z2-7]{55}$"
          description: Stellar public key (56 characters, starting with "G").
          example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
        subscription:
          $ref: "#/components/schemas/PushSubscription"
        periods:
          $ref: "#/components/schemas/PeriodPrefs"
    EmailSubscribeRequest:
      type: object
      required: [walletAddress, email, periods]
      properties:
        walletAddress:
          type: string
          pattern: "^G[A-Z2-7]{55}$"
          description: Stellar public key (56 characters, starting with "G").
          example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
        email:
          type: string
          format: email
          example: "alice@example.com"
        periods:
          $ref: "#/components/schemas/PeriodPrefs"
    UnsubscribeRequest:
      type: object
      description: One of `token` or `walletAddress` + `channel` (mutually exclusive).
      properties:
        token:
          type: string
          description: Unsubscribe token from a notification email.
          example: "f3a5c1d9e2b8476fa0b1c2d3e4f5a6b7"
        walletAddress:
          type: string
          pattern: "^G[A-Z2-7]{55}$"
          description: Stellar public key.
          example: "GABQUEIYD4TC2NB3IJEVAV26MVWHG6UBRCHZNHNEVOZLTQGHZ3K5ZEJS"
        channel:
          type: string
          enum: [push, email]
          description: Channel to disable on the preference page.
          example: "email"
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string